fix overflow error

+++ perfrunner/settings.py
@@ -239,7 +239,7 @@ class PhaseSettings(object):
         self.reads = int(options.get('reads', self.READS))
         self.updates = int(options.get('updates', self.UPDATES))
         self.deletes = int(options.get('deletes', self.DELETES))
-        self.ops = int(options.get('ops', self.OPS))
+        self.ops = float(options.get('ops', self.OPS))
         self.throughput = float(options.get('throughput', self.THROUGHPUT))
 
         self.size = int(options.get('size', self.SIZE))
